Fix bug #763: Problems with IceTV fetch when IceTV is disabled
Only show the "Fetch EPG and update timers now" MENU>IceTV menu
entry when IceTV is enabled.

There is no interface in ChoiceBox (which IceTVMain inherits from)
to change the ChoiceBox menu list, so a new instance of IceTVMain
is created when the enable/disable status is changed and the old
instance is closed.

The default parameter selection is used to pass the old selection
position to the new instance so that thbe selected position is
maintained acrtoss instances of IceTVMain.

The old instance of IceTVMain is hidden before the new one is created -
this prevents old instances of IceTVMain on the dialog stack being
briefly displayed when the top instance is closed.

Move the "Fetch EPG and update timers now" to the end of the menu
list where its appearance and removal will be less visually disruptive.
"Fetch EPG and update timers now" vcan be reached with a single UP
from the top of the list (it was previously reached with a single
DOWN.

Move the "IceTV setup wizard" entry after the "Login to IceTV server"
entry, because the former is used less frequently than the latter.