[Scanner] Fix bug #765: File scanner lists include macOS Finder metadata files
Filter out filenames that start with "._" (macOS Finder metadata
files) when when scanning a device for files of a particular type.

Because the metadata files have the same extension as the actual
file, they are selected by Scanners even though the metadata file
has unrelated content to what the extension suggests.

This kind of filtering is already done for file list creation in
Components.MovieList.