Fix issue #768: [IceTV] Improve portability as plugin
[IceTV]

The IceTV plugin depends on definitions of epgclock_icetv.png and
timer_icetv.png in the skin, and requires the system ActionMap
context DirectionActions to contain the non-standard chplus and
chminus actions (they were added for the IceTV plugin).

Make the IceTV plugin more portable by having local copies of the
two pixmap files, and check for the pixmaps in the skin first, then
in the plugin, and by having a local keymap.xml file to support the
chplus/chminus actions for IceTV debug logging control.

Now that the chplus/chminus actions are available for the plugin
in its local keymap.xml, they could be removed from the system-wide
keymap.xml if they are not referenced elsewhere.